Add files via upload
We are given a sensor network with a fairly small number of MAC addresses, one for each sensor, and we
would like to group the sensors to form several cluster of sensors based on their MAC address. A MAC
address (Media Access Control address) is a unique identifier assigned to a network interface controller (NIC)
for on a network. A MAC Address is often referred to as a hardware address. The MAC addresses are unique
and they are used for communication between components or computers. MAC addresses take the form of
alphanumeric numbers in hexadecimal format arranged in groups of 2 (usually) separated by a colon.
Renaming MAC addresses to minimize bits used for communication
(https://dl.acm.org/doi/abs/10.1145/501449.501463 ) is one technique used to work around MAC addresses
that are unique and drawn from an extremely large number of possible values.